requirements.txt
annotated-types==0.7.0
asttokens==2.4.1
azure-core==1.30.2
azure-data-tables==12.5.0
azure-storage-blob==12.22.0
beautifulsoup4==4.12.3
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
cryptography==43.0.0
debugpy==1.8.5
decorator==5.1.1
ensure==1.0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.0.1
frozendict==2.4.4
html5lib==1.1
idna==3.7
importlib_metadata==8.2.0
# Editable install with no version control (Ingest==0.0.0)
-e .
ipykernel==6.29.5
ipython==8.18.1
isodate==0.6.1
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
lxml==5.3.0
matplotlib-inline==0.1.7
multidict==6.0.5
multitasking==0.0.11
nest-asyncio==1.6.0
numpy==2.0.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
parso==0.8.4
peewee==3.17.6
platformdirs==4.2.2
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
python-box==7.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pywin32==306
PyYAML==6.0.2
pyzmq==26.1.1
requests==2.32.3
six==1.16.0
soupsieve==2.6
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
yarl==1.9.4
yfinance==0.2.41
zipp==3.20.0
